Fixes for test framework compatibility

+++ VortexEngine/src/Buttons/Button.cpp
@@ -59,7 +59,7 @@ void Button::check()
 
   // read the new button state
 #ifdef VORTEX_LIB
-  uint8_t newButtonState = digitalRead(9);
+  uint8_t newButtonState = (uint8_t)digitalRead(9);
 #elif defined(VORTEX_ARDUINO)
   uint8_t newButtonState = (PORTB.IN & PIN2_bm) ? 1 : 0;
 #endif
